Move libutil internal headers out of include path
