+++ native/jni/utils/misc.h
